modelo/requirements-minimal.txt
onnxruntime
ray[serve]
fastapi
uvicorn
python-multipart
Pillow
numpy